backend/requirements.txt
# Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Base de datos
pymongo==4.6.1
motor==3.3.2
beanie==1.25.0

# Autenticación
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Caché
redis==5.0.1

# Validación
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
